+++ src/layouts/SidebarLayout.styles.ts
@@ -5,8 +5,8 @@ export const styles = css`
     position: relative;
     display: flex;
     flex-direction: row;
-    width: 100vw;
-    height: 100vh;
+    width: 100%;
+    height: 100%;
     overflow: hidden;
   }