+++ src/index.ts
@@ -1,8 +1,7 @@
-import { app } from './app';
-
+import { app } from './App.js';
 export type * from './types/AppConfigs.js';
 export type * from './types/AppOptions.js';
 export type * from './types/AppTypes.js';
 
-export { app } from './app';
+export { app };
 export default app;